Support canfd (#1)
* Updated Message.Length to uint16 to support message lengths up to 65,535 bytes

* Updated Signal.Lenght and Signal.Size to uint16 for payloads larger than 8 bytes

* Added payload.go and payload_tests.go

* Updated Big and Little Endian encoding methods

* Added benchmarks

* Added methods and tests to decode signed/unsigned signals in big/little endian

* Added benchmarks for UnsignedBitsLittleEndian

* Added hex string methods for Payload

* Added UnmarshalPhysicalPayload and updated bit-level methods for Payload struct

* Added UnmarshalValueDescriptionPayload and associated methods for Payload data struct

* Added decode_tests.go

* Updated decode test

* Added asserts to decode_test.go

* Added comparitive unit tests for can.Data and benchmarks for decoding

* Updated PackLittleEndian and PackBigEndian to return new big.Int and reference pointer to payload in decode methods

* Fixed units in VDM_Disconnect test

+++ frame_json_test.go
@@ -102,7 +102,7 @@ func (Frame) Generate(rand *rand.Rand, size int) reflect.Value {
 	} else {
 		f.ID = rand.Uint32() & MaxID
 	}
-	f.Length = uint8(rand.Intn(9))
+	f.Length = uint16(rand.Intn(9))
 	if !f.IsRemote {
 		_, _ = rand.Read(f.Data[:f.Length])
 	}
